put content back to scoreboard.js, fixed data reset on reload, 	rendering output onclick
