+++ game/game.html
@@ -76,5 +76,6 @@
 
     <script src="./game.js"></script>
     <script src="./inst-game.js"></script>
+    <script src="./scoreboardLogic.js"></script>
   </body>
 </html>
