+++ game/scoreboard.js
@@ -1,4 +1,4 @@
-export const renderScore = (data, parentEl) => {
+const renderScore = (data, parentEl) => {
   const listContainer = document.createElement("ul");
   data.map(item => {
     const domListItem = document.createElement("li");
@@ -12,3 +12,69 @@ export const renderScore = (data, parentEl) => {
   });
   parentEl.appendChild(listContainer);
 };
+
+
+const btnx = document.getElementById('scoreboardBtn')
+const scoreboard = document.getElementById('scoreboard')
+
+const resetScore = () => localStorage.setItem("gameScores", JSON.stringify([]));
+const getScore = JSON.parse(localStorage.getItem("gameScores")) || resetScore();
+
+const setScore = data => localStorage.setItem("gameScores", JSON.stringify(data));
+const addScore = (player, score) => {
+  const newScoreData = {
+    name: player,
+    score: score
+  };
+  getScore.push(newScoreData);
+  return setScore(getScore);
+};
+
+const generateRandomScore = () => Math.floor(Math.random() * 1000);
+const fillScores = num => {
+  const nameArr = [
+    "Kent Sauro",
+    "Donald Mitten",
+    "Sherill Mayson",
+    "Wilbur Overton",
+    "Alison Daughtrey",
+    "Coleen Carrington",
+    "Mara Royce",
+    "Vickey Sarris",
+    "Patsy Narducci",
+    "Josue Weitzel"
+  ];
+  for (let i = 0; i < num; i++) {
+    addScore(nameArr[i], generateRandomScore());
+  }
+};
+
+const sortScores = arr => {
+  const sortedScores = arr.sort((a, b) => {
+    a.score < b.score ? 1 : -1;
+  });
+  return sortedScores;
+};
+
+const checkScores = arr => {
+  if (arr.length) {
+    let output = ''
+    if (arr.length <= 10) {
+      output = sortScores(arr);
+      return renderScore(output, scoreboard)
+      //push score using addScore function
+    } else {
+      const tenScoresArr = sortScores(arr).slice(0, 10);
+      output = setScore(tenScoresArr)
+      console.log("tenScoresArr updated");
+      renderScore(output, scoreboard)
+      //save scores in localStorage
+      //render on scoreboard
+    }
+  }
+};
+
+btnx.addEventListener('click' , () => {
+  console.log('dupa1')
+  checkScores(getScore)
+})
